requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
joblib==1.3.2
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
python-multipart==0.0.9
scikit-fuzzy==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
sniffio==1.3.1
starlette==0.36.3
threadpoolctl==3.3.0
typing_extensions==4.10.0
uvicorn==0.27.1
wheel==0.42.0
